Move witai and weather models to a different directory
