+++ app/src/main/java/kapil/voiceassistedweatherapp/weather/models/witai/Location.java
@@ -1,5 +1,5 @@
 
-package kapil.voiceassistedweatherapp.witai.models;
+package kapil.voiceassistedweatherapp.weather.models.witai;
 
 import com.fasterxml.jackson.annotation.JsonAnyGetter;
 import com.fasterxml.jackson.annotation.JsonAnySetter;
@@ -15,9 +15,10 @@ import java.util.Map;
 @JsonPropertyOrder({
     "confidence",
     "type",
-    "value"
+    "value",
+    "suggested"
 })
-public class Intent {
+public class Location {
 
     @JsonProperty("confidence")
     private Double confidence;
@@ -25,6 +26,8 @@ public class Intent {
     private String type;
     @JsonProperty("value")
     private String value;
+    @JsonProperty("suggested")
+    private Boolean suggested;
     @JsonIgnore
     private Map<String, Object> additionalProperties = new HashMap<String, Object>();
 
@@ -56,6 +59,16 @@ public class Intent {
     @JsonProperty("value")
     public void setValue(String value) {
         this.value = value;
+    }
+
+    @JsonProperty("suggested")
+    public Boolean getSuggested() {
+        return suggested;
+    }
+
+    @JsonProperty("suggested")
+    public void setSuggested(Boolean suggested) {
+        this.suggested = suggested;
     }
 
     @JsonAnyGetter
